+++ src/main/java/starlight/backend/proof/model/entity/ProofEntity.java
@@ -6,14 +6,13 @@ import jakarta.validation.constraints.NotBlank;
 import lombok.*;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.URL;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.validation.annotation.Validated;
 import starlight.backend.proof.model.enums.Status;
 import starlight.backend.skill.model.entity.SkillEntity;
 import starlight.backend.user.model.entity.UserEntity;
 
 import java.time.Instant;
-import java.util.Set;
+import java.util.List;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
 
@@ -55,7 +54,8 @@ public class ProofEntity {
             joinColumns = @JoinColumn(name = "proof_id"),
             inverseJoinColumns = @JoinColumn(name = "skill_id"))
     @JsonManagedReference
-    private Set<SkillEntity> skills;
+    private List<SkillEntity> skills;
+
     public void setDateLastUpdated(Instant now) {
         this.dateLastUpdated = now;
     }
