+++ src/main/java/starlight/backend/proof/model/response/ProofFullInfoWithSkills.java
@@ -3,6 +3,7 @@ package starlight.backend.proof.model.response;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
 import starlight.backend.proof.model.enums.Status;
+import starlight.backend.skill.model.response.SkillWithCategory;
 
 import java.time.Instant;
 import java.util.LinkedList;
