--- src/main/java/starlight/backend/proof/model/response/SkillWithCategory.java
@@ -1,9 +1,0 @@
-package starlight.backend.proof.model.response;
-
-import lombok.Builder;
-
-@Builder
-public record SkillWithCategory(
-       String skill,
-       String category
-) {}
