+++ src/main/java/starlight/backend/skill/SkillController.java
@@ -17,7 +17,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import starlight.backend.skill.model.request.AddSkill;
-import starlight.backend.skill.model.response.ProofWithSkills;
+import starlight.backend.proof.model.response.ProofWithSkills;
 import starlight.backend.skill.model.response.SkillList;
 import starlight.backend.skill.model.response.SkillListWithPagination;
 import starlight.backend.skill.service.SkillServiceInterface;
