+++ src/main/java/starlight/backend/skill/SkillMapper.java
@@ -2,10 +2,11 @@ package starlight.backend.skill;
 
 import org.mapstruct.Mapper;
 import starlight.backend.proof.model.entity.ProofEntity;
+import starlight.backend.proof.model.response.ProofWithSkills;
 import starlight.backend.skill.model.entity.SkillEntity;
-import starlight.backend.skill.model.response.ProofWithSkills;
 import starlight.backend.skill.model.response.SkillList;
 import starlight.backend.skill.model.response.SkillListWithPagination;
+import starlight.backend.skill.model.response.SkillWithCategory;
 
 import java.util.List;
 
@@ -15,14 +16,26 @@ import static org.mapstruct.ReportingPolicy.IGNORE;
 public interface SkillMapper {
     default SkillListWithPagination toSkillListWithPagination(List<SkillEntity> skills, long pageNumber) {
         return SkillListWithPagination.builder()
-                .data(skills)
+                .data(skills.stream()
+                        .map(this::toSkillWithCategory)
+                        .toList())
                 .total(pageNumber)
                 .build();
     }
 
     default SkillList toSkillList(List<SkillEntity> skills) {
         return SkillList.builder()
-                .skills(skills)
+                .skills(skills.stream()
+                        .map(this::toSkillWithCategory)
+                        .toList())
+                .build();
+    }
+
+    default SkillWithCategory toSkillWithCategory(SkillEntity skill) {
+        return SkillWithCategory.builder()
+                .skillId(skill.getSkillId())
+                .category(skill.getCategory())
+                .skill(skill.getSkill())
                 .build();
     }
