+++ src/main/java/starlight/backend/skill/model/response/SkillList.java
@@ -1,12 +1,11 @@
 package starlight.backend.skill.model.response;
 
 import lombok.Builder;
-import starlight.backend.skill.model.entity.SkillEntity;
 
 import java.util.List;
 
 @Builder
 public record SkillList(
-        List<SkillEntity> skills
+        List<SkillWithCategory> skills
 ) {
 }
