+++ src/main/java/starlight/backend/skill/model/response/SkillWithCategory.java
@@ -1,0 +1,11 @@
+package starlight.backend.skill.model.response;
+
+import lombok.Builder;
+
+@Builder
+public record SkillWithCategory(
+        long skillId,
+        String skill,
+        String category
+) {
+}
