+++ src/main/java/starlight/backend/skill/service/SkillServiceInterface.java
@@ -1,8 +1,8 @@
 package starlight.backend.skill.service;
 
 import org.springframework.security.core.Authentication;
+import starlight.backend.proof.model.response.ProofWithSkills;
 import starlight.backend.skill.model.request.AddSkill;
-import starlight.backend.skill.model.response.ProofWithSkills;
 import starlight.backend.skill.model.response.SkillList;
 import starlight.backend.skill.model.response.SkillListWithPagination;
